Perbaikan mekanisme untuk multiple callback pada ArchPresenter
Mekanisme multiple callback pada ArchPresenter berubah menjadi menggunakan data class PresenterRequest agar Request dan Callback menjadi satu-kesatuan.

+++ app/src/main/java/sidev/lib/implementation/_simulation/sigudang/fragments/BoundProcessFrag/BoundWarehouseSelect.kt
@@ -66,7 +66,7 @@ open class BoundWarehouseSelect : Frag(), BoundProcessRootFrag, TopMiddleBottomB
 
     @CallSuper
     override fun _initView(layoutView: View) {
-        layoutView.setPadding(10.dp)
+        layoutView.setPadding(10.dp.toInt())
         bsWarehouseSelect= BsWarehouseSelectFr()
         bsWarehouseSelect.setTitle("Pilih Gudang")
         bsWarehouseSelect.setBtnConfirmText("Pilih")
